On stm32F2xx gpio is encoded to 20bits but the button definitions use 16bit type.

+++ configs/olimex-stm32-h405/src/stm32_buttons.c
@@ -65,7 +65,7 @@
  * the BUTTON_* definitions in board.h
  */
 
-static const uint16_t g_buttons[NUM_BUTTONS] =
+static const uint32_t g_buttons[NUM_BUTTONS] =
 {
   GPIO_BTN_BUT
 };
